fixed vector data type from int to long long int

+++ include/Alg1.h
@@ -5,8 +5,8 @@
 
 using namespace std;
 
-pair<int, int> Alg1( int* vecX, int* vecY, bool** resultMtx, int n );
+pair<int, int> Alg1( long long int* vecX, long long int* vecY, bool** resultMtx, int n );
 
-pair<int, int> Alg1_Prob( int* vecX, int* vecY, int n, const int PROB );
+pair<int, int> Alg1_Prob( long long int* vecX, long long int* vecY, int n, const int PROB );
 
 #endif
